react-twemojiの型定義替え
